requirements.txt
# Core
streamlit>=1.32.0
pandas>=2.0.0
httpx>=0.25.0

# Databricks
databricks-sdk>=0.20.0

# LLM clients
anthropic>=0.18.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
